Resolve merge conflict by incorporating both suggestions

+++ nym-validator-rewarder/src/rewarder/storage/manager.rs
@@ -114,9 +114,9 @@ impl StorageManager {
     pub(crate) async fn insert_rewarding_epoch_credential_issuance(
         &self,
         epoch: i64,
-        starting_dkg_epoch: u32,
-        ending_dkg_epoch: u32,
-        total_issued_partial_credentials: u32,
+        starting_dkg_epoch: i64,
+        ending_dkg_epoch: i64,
+        total_issued_partial_credentials: i64,
         budget: String,
     ) -> Result<(), sqlx::Error> {
         sqlx::query!(
